ecom_airflow/requirements.txt
# Astro Runtime includes the following pre-installed providers packages: https://www.astronomer.io/docs/astro/runtime-image-architecture#provider-packages
# Core Airflow and Providers
astronomer-cosmos
apache-airflow-providers-snowflake
apache-airflow-providers-postgres
apache-airflow-providers-amazon
faker==19.3.0
boto3==1.35.77
psycopg2-binary==2.9.10
python-dotenv==1.0.1
SQLAlchemy
pyarrow==18.1.0
snowflake-connector-python==3.12.4